None
for example we're just starting a video game and
in this video game we have a user's weapons but
when you're just turning the game there are no weapons
Well in that case we might want to assign the variable weapons
maybe in our code or we use this variable
but we also want to know that this user doesn't really have any weapons.
so we use None.